重构搜索注解
