+++ APDPlat_Core/src/main/java/org/apdplat/module/dictionary/model/DicItem.java
@@ -30,9 +30,9 @@ import javax.xml.bind.annotation.XmlTransient;
 import javax.xml.bind.annotation.XmlType;
 import org.apdplat.platform.annotation.Database;
 import org.apdplat.platform.model.SimpleModel;
-import org.apdplat.platform.search.Searchable;
-import org.apdplat.platform.search.SearchableComponent;
-import org.apdplat.platform.search.SearchableProperty;
+import org.apdplat.platform.search.annotations.Searchable;
+import org.apdplat.platform.search.annotations.SearchableComponent;
+import org.apdplat.platform.search.annotations.SearchableProperty;
 import org.hibernate.annotations.Cache;
 import org.hibernate.annotations.CacheConcurrencyStrategy;
 import org.springframework.context.annotation.Scope;
