+++ APDPlat_Core/src/main/java/org/apdplat/module/monitor/model/ProcessTime.java
@@ -33,9 +33,9 @@ import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
 import org.apdplat.platform.annotation.Database;
 import org.apdplat.platform.model.Model;
-import org.apdplat.platform.search.Index;
-import org.apdplat.platform.search.Searchable;
-import org.apdplat.platform.search.SearchableProperty;
+import org.apdplat.platform.search.annotations.Index;
+import org.apdplat.platform.search.annotations.Searchable;
+import org.apdplat.platform.search.annotations.SearchableProperty;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
