+++ APDPlat_Core/src/main/java/org/apdplat/platform/model/Model.java
@@ -51,9 +51,9 @@ import org.apdplat.platform.annotation.ModelCollRef;
 import org.apdplat.platform.annotation.RenderDate;
 import org.apdplat.platform.annotation.RenderTime;
 import org.apdplat.platform.log.APDPlatLoggerFactory;
-import org.apdplat.platform.search.SearchableComponent;
-import org.apdplat.platform.search.SearchableId;
-import org.apdplat.platform.search.SearchableProperty;
+import org.apdplat.platform.search.annotations.SearchableComponent;
+import org.apdplat.platform.search.annotations.SearchableId;
+import org.apdplat.platform.search.annotations.SearchableProperty;
 
 /**
  *
