+++ APDPlat_Core/src/main/java/org/apdplat/platform/search/annotations/Store.java
@@ -18,7 +18,7 @@
  *
  */
 
-package org.apdplat.platform.search;
+package org.apdplat.platform.search.annotations;
 
 /**
  *
