+++ APDPlat_Module/src/main/java/org/apdplat/module/info/model/News.java
@@ -34,8 +34,8 @@ import javax.persistence.OneToMany;
 import javax.persistence.Transient;
 import org.apdplat.platform.annotation.Database;
 import org.apdplat.platform.model.SimpleModel;
-import org.apdplat.platform.search.Searchable;
-import org.apdplat.platform.search.SearchableComponent;
+import org.apdplat.platform.search.annotations.Searchable;
+import org.apdplat.platform.search.annotations.SearchableComponent;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
