+++ APDPlat_Module/src/main/java/org/apdplat/module/info/model/NewsContent.java
@@ -29,8 +29,8 @@ import javax.persistence.Lob;
 import javax.persistence.ManyToOne;
 import org.apdplat.platform.annotation.Database;
 import org.apdplat.platform.model.SimpleModel;
-import org.apdplat.platform.search.Searchable;
-import org.apdplat.platform.search.SearchableProperty;
+import org.apdplat.platform.search.annotations.Searchable;
+import org.apdplat.platform.search.annotations.SearchableProperty;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
